+++ README.md
@@ -22,7 +22,7 @@ There are a significant number of controls in ASP.NET Web Forms, and we will foc
 
   - Editor Controls
 	  - [Button](docs/Button.md)
-    - LinkButton
+    - [LinkButton](docs/LinkButton.md)
   - Data Controls
     - Chart(?)
     - [DataList](docs/DataList.md)
