+++ src/BlazorWebFormsComponents/LinkButton.razor
@@ -1,0 +1,10 @@
+﻿@inherits ButtonBaseComponent
+
+@if (PostBackUrl == null)
+{
+	<a id="@ID" @onclick="@Click">@Text</a>
+}
+else
+{
+	<a id="@ID" @onclick="@Click" href="@PostBackUrl">@Text</a>
+}
