+++ src/BlazorWebFormsComponents/LinkButton.razor.cs
@@ -1,0 +1,7 @@
+﻿namespace BlazorWebFormsComponents
+{
+	public partial class LinkButton : ButtonBaseComponent
+	{
+
+	}
+}
